Resize table and add Pink and Orange hands
Resize the table to allow for adding the 5th and 6th players on the side
of the board.

I could put them next to the current players, but then the new players
would be further from the main board. I like them facing the side of the
board just fine. That's how we do it in real life!