Place barbarian tokens under other tokens
With the addition of barbarian tokens at the home provences there really
isn't enough room for the barbarian tokens, goods tokens and food
tokens. The example in the rulebook for section "0.8 Distribute
Barbarian Tokens" clearly shows a food token on top of a barbarian token
and I think that's a good idea since the barbarian tokens are larger in
size and will not be completely covered by other tokens.