Update maps to production version
Update to use the production version of the map. Ensure that all
boundries match those of the v1.4 version of the rulebook.

Since the map has a bit more boarder to it, the scale changed a little
and most things need to be nudged.

I've had to move all of the boundry markers, so while I was at it I
pushed them inside the out-of-bounds area to provide the most in-bounds
map usage area possible.

Other things to be nudged in the following commits.