Replace discs with custom cylinder models
The White Backgammon Piece that was being used for all of the marker
discs has a square collider box. This was making it so the pieces would
bump against each other when place on snap points near each other on the
Improvement Track.

The model itself is just reused from the custom bags for the LG, MG
markers.

While I was at it, I added a texture to the custom disc model. It's a
picutre I personally took so it may not be the best texture in the
world, but there are no copywrite issues or anything to deal with.

The disc is a little darker, but personally I like the change. It
actually looks wooden as it should (the backgammon piece is clearly
plastic or ceramic).