Make the Barbarian / Militia Cube bag infinite
It's much easier to clean up the battle board by deleting grey cubes
than it is to return them to the bag. Since the limit is on how many
there should be in a given battle, and not how many there should be in
the game, there's no reason to not make these infinite.